chore: cleanup and notes
