+++ packages/console/src/components/Breadcrumbs/async/fn.ts
@@ -2,30 +2,51 @@ import { listWorkflows } from 'models/Workflow/api';
 import { listNamedEntities } from 'models/Common/api';
 import { ResourceType, defaultPaginationConfig } from 'models';
 import { listProjects } from 'models/Project/api';
-import { Routes } from 'routes';
-import startCase from 'lodash/startCase';
-import camelCase from 'lodash/camelCase';
 import { listExecutions } from 'models/Execution/api';
 import {
   formatEntities,
   formatProjectEntities,
   formatProjectEntitiesAsDomains,
 } from './utils';
-import { Breadcrumb, BreadcrumbEntity } from '../types';
-import { namedEntitiesUrlSegments } from '../validators';
+import { namedEntitiesList } from '../defaultValue';
 
+/**
+ * Default async data function, returns an empty array
+ * Used for breadcumb initialization
+ * @param _projectId
+ * @param _domainId
+ * @returns
+ */
 export const defaultVoid = async (_projectId = '', _domainId = '') => [];
 
+/**
+ * Fetch a list of projects and format them for the breadcrumb
+ * @param _projectId
+ * @param _domainId
+ * @returns
+ */
 export const projects = async (_projectId = '', _domainId = '') => {
   return listProjects().then(data => formatProjectEntities(data));
 };
 
+/**
+ * Fetch a list of domains and format them for the breadcrumb
+ * @param projectId
+ * @param domainId
+ * @returns
+ */
 export const domains = async (projectId = '', domainId = '') => {
   return listProjects().then(data =>
     formatProjectEntitiesAsDomains(data, projectId, domainId),
   );
 };
 
+/**
+ * Fetch a list of workflows and format them for the breadcrumb
+ * @param projectId
+ * @param domainId
+ * @returns
+ */
 export const workflows = async (projectId = '', domainId = '') => {
   return listWorkflows({
     project: projectId,
@@ -33,6 +54,12 @@ export const workflows = async (projectId = '', domainId = '') => {
   }).then(data => formatEntities(data));
 };
 
+/**
+ * Fetch a list of tasks and format them for the breadcrumb
+ * @param projectId
+ * @param domainId
+ * @returns
+ */
 export const tasks = async (projectId = '', domainId = '') => {
   return listNamedEntities(
     {
@@ -44,6 +71,12 @@ export const tasks = async (projectId = '', domainId = '') => {
   ).then(data => formatEntities(data));
 };
 
+/**
+ * Fetch a list of launch plans and format them for the breadcrumb
+ * @param projectId
+ * @param domainId
+ * @returns
+ */
 export const launchPlans = async (projectId = '', domainId = '') => {
   return listNamedEntities(
     {
@@ -55,52 +88,18 @@ export const launchPlans = async (projectId = '', domainId = '') => {
   ).then(data => formatEntities(data));
 };
 
-export const namedEntitiesList = (projectId = '', domainId = '') => {
-  const workflow = {
-    title: 'Workflows',
-    createdAt: '',
-    url: Routes.WorkflowDetails.makeUrl(projectId, domainId, ''),
-  };
-  const task = {
-    title: 'Tasks',
-    createdAt: '',
-    url: Routes.TaskDetails.makeUrl(projectId, domainId, ''),
-  };
-  const launchPlans = {
-    title: 'Launch Plans',
-    createdAt: '',
-    url: Routes.TaskDetails.makeUrl(projectId, domainId, ''),
-  };
-
-  return [workflow, task, launchPlans];
-};
-
+/**
+ * Returns a list of named entities (workflows, tasks, launch plans).
+ * Preformatted for the breadcrumb popover.
+ * @param projectId
+ * @param domainId
+ * @returns
+ */
 export const namedEntities = async (projectId = '', domainId = '') => {
   return namedEntitiesList(projectId, domainId);
 };
 
-export const namedEntitiesDefaultValue = (
-  location: Location,
-  _breadcrumb: Breadcrumb,
-) => {
-  const segments = location.pathname.split('/');
-  const namedEntitySegment =
-    namedEntitiesUrlSegments.find(e => segments.find(s => s === e)) || '';
-
-  const normalizedNamedEntitySegment = namedEntitySegment.endsWith('s')
-    ? camelCase(namedEntitySegment)
-    : `${camelCase(namedEntitySegment)}s`;
-
-  const namedEntitiesBreadcumbPopOverList: BreadcrumbEntity[] =
-    namedEntitiesList('', '');
-  const titles = namedEntitiesBreadcumbPopOverList.map(entity =>
-    camelCase(entity.title),
-  );
-  const entity =
-    titles.find(title => title === normalizedNamedEntitySegment) || '';
-  return startCase(entity);
-};
-
+// TODO: Split this into a seperate lookup function per version type
 export const namedEntitiesVersions = async (projectId = '', domainId = '') => {
   const segments = decodeURI(window.location.pathname).split('/');
   const versionIndex = segments.findIndex(segment => segment === 'version');
@@ -111,26 +110,4 @@ export const namedEntitiesVersions = async (projectId = '', domainId = '') => {
     domain: domainId,
     name,
   }).then(data => formatEntities(data));
-};
-
-export const namedEntitiesVersionsViewAll = (projectId = '', domainId = '') => {
-  // const segments = decodeURI(window.location.pathname).split('/');
-  // const versionIndex = segments.findIndex(segment => segment === 'version');
-  // const nameIndex = versionIndex - 2;
-  // const name = segments[nameIndex];
-
-  /// TODO: namedEntitiesUrlSegments
-
-  // const routesKeys = Object.keys(Routes.ProjectDetails.sections);
-  // const routesKey = routesKeys.find(key => key.includes(name)) || '';
-  // const routeSection = Routes.ProjectDetails.sections[routesKey];
-  // const makeUrl =
-  //   typeof routeSection['makeUrl'] !== 'undefined' &&
-  //   typeof routeSection.makeUrl === 'function'
-  //     ? routeSection.makeUrl
-  //     : Routes.ProjectDashboard.makeUrl;
-
-  // return makeUrl(projectId, domainId, name);
-
-  return '/todo';
 };
