+++ packages/console/src/components/Breadcrumbs/defaultValue/default.ts
@@ -1,0 +1,1 @@
+export const defaultValue = '';
